// astros_vue/src/enums/scripts/humanCyborgRelations.ts
export enum HumanCyborgRelationsCmd {
  MILD_HAPPY = 1,
  EXTREME_HAPPY = 2,
  MILD_SAD = 3,
  EXTREME_SAD = 4,
  MILD_ANGRY = 5,
  EXTREME_ANGRY = 6,
  MILD_SCARED = 7,
  EXTREME_SCARED = 8,
  OVERLOAD = 9,
  ENABLE_MUSE = 10,
  DISABLE_MUSE = 11,
  TOGGLE_MUSE = 12,
  TRIGGER_MUSING = 13,
  MIN_SECONDS_BETWEEN_MUSINGS = 14,
  MAX_SECONDS_BETWEEN_MUSINGS = 15,
  PLAY_WAV_ON_A = 16,
  PLAY_WAV_ON_B = 17,
  PLAY_SD_RANDOM_ON_A = 18,
  PLAY_SD_RANDOM_ON_B = 19,
  PANIC_STOP = 20,
  GRACEFUL_STOP = 21,
  STOP_WAV_ON_A = 22,
  STOP_WAV_ON_B = 23,
  VOCALIZER_VOLUME = 24,
  WAV_A_VOLUME = 25,
  WAV_B_VOLUME = 26,
  ENABLE_IMPROV = 27,
  ENABLE_CANONICAL = 28,
  ENABLE_PERSONALITY_OVERRIDE = 29,
  DISABLE_PERSONALITY_OVERRIDE = 30,
  ZERO_EMOTIONS = 31,
  SET_HAPPY_LEVEL = 32,
  SET_SAD_LEVEL = 33,
  SET_MAD_LEVEL = 34,
  SET_SCARED_LEVEL = 35,
}

export enum HcrCommandCategory {
  NONE,
  STIMULI,
  MUSE,
  SD_WAV,
  STOP,
  VOLUME,
  OVERRIDE,
}